requirements.txt
SQLAlchemy >= 2.0.20
psycopg2-binary >= 2.9.6
Flask >= 2.3.2
Werkzeug >= 2.3.6
flask-login >= 0.6.2
Jinja2 >= 3.1.2
WTForms >= 3.0.1
weasyprint >= 59.0
python-dotenv >= 1.0.
waitress >= 2.1.2
passlib >= 1.7.4
PyPDF4 >= 1.27.0